Add test to ensure tera templates are in sync
